requirements.txt
django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-celery==3.3.1
google-api-python-client==2.58.0
